13_rpn_calculator specs are now satisfied
The code does work but is not very dry.  Some refactoring will
defintely be done in the near future.
The code passing the final 2 tests most likely does not fulfill the
implied functionality.  It does however pass the specs in accordance
to the discipline of writing the simplest code that will work.